+++ CHANGELOG.md
@@ -11,6 +11,8 @@
 * Replace `chat` lens on `Has Update Chat` instance
 
 # 0.1.3
+* Define `Has Callback Chat` instance
+* Define `Has Callback From` instance
 * Define `Has Member From` instance
 * Define `Has Message Chat` instance
 * Define `Has Message From` instance